fix: postpon the call to the server
