+++ pages/main/student/cross-check-submit/index.tsx
@@ -306,43 +306,43 @@ const CrossCheckSubmitPage: React.FC<PropsCrossCheckPage> = ({
 
 export default CrossCheckSubmitPage;
 
-export const getServerSideProps = async () => {
-  let tasksData: ITask[] = [] as ITask[];
-  await db
-    .collection('TasksArray')
-    .get()
-    .then((snap) => {
-      if (snap !== undefined && snap !== null) {
-        tasksData = snap.docs.map((doc) => doc.data()) as ITask[];
-      } else {
-        tasksData = [] as ITask[];
-      }
-    });
-  let courseData: ICourse[] = [] as ICourse[];
-  await db
-    .collection('sessions')
-    .get()
-    .then((snap) => {
-      if (snap !== undefined && snap !== null) {
-        courseData = snap.docs.map((doc) => doc.data()) as ICourse[];
-      } else {
-        courseData = [] as ICourse[];
-      }
-    });
+// export const getServerSideProps = async () => {
+//   let tasksData: ITask[] = [] as ITask[];
+//   await db
+//     .collection('TasksArray')
+//     .get()
+//     .then((snap) => {
+//       if (snap !== undefined && snap !== null) {
+//         tasksData = snap.docs.map((doc) => doc.data()) as ITask[];
+//       } else {
+//         tasksData = [] as ITask[];
+//       }
+//     });
+//   let courseData: ICourse[] = [] as ICourse[];
+//   await db
+//     .collection('sessions')
+//     .get()
+//     .then((snap) => {
+//       if (snap !== undefined && snap !== null) {
+//         courseData = snap.docs.map((doc) => doc.data()) as ICourse[];
+//       } else {
+//         courseData = [] as ICourse[];
+//       }
+//     });
 
-  let completedTasksData: IWorkDone[] = [] as IWorkDone[];
-  await db
-    .collection('completed_tasks')
-    .get()
-    .then((snap) => {
-      if (snap !== undefined && snap !== null) {
-        completedTasksData = snap.docs.map((doc) => doc.data()) as IWorkDone[];
-      } else {
-        completedTasksData = [] as IWorkDone[];
-      }
-    });
+//   let completedTasksData: IWorkDone[] = [] as IWorkDone[];
+//   await db
+//     .collection('completed_tasks')
+//     .get()
+//     .then((snap) => {
+//       if (snap !== undefined && snap !== null) {
+//         completedTasksData = snap.docs.map((doc) => doc.data()) as IWorkDone[];
+//       } else {
+//         completedTasksData = [] as IWorkDone[];
+//       }
+//     });
 
-  return {
-    props: { tasksData, courseData, completedTasksData },
-  };
-};
+//   return {
+//     props: { tasksData, courseData, completedTasksData },
+//   };
+// };
