+++ pages/main/student/index.tsx
@@ -5,8 +5,18 @@ import { List } from 'antd';
 import StudentHome from '../../../components/student/student-houm/StudentHome';
 import CrossCheckSubmitPage from './cross-check-submit';
 import CrossCheckReviewPage from './cross-check-review';
+import { db } from '../../../firebase';
+import { ITask } from '../../../interfaces/ITask';
+import { ICourse } from '../../../interfaces/ICourse';
+import { IWorkDone } from '../../../interfaces/IWorkDone';
 
-const StudentPage: React.FC = () => {
+interface PropsStudentPage {
+  tasksData: ITask[];
+  courseData: ICourse[];
+  completedTasksData: IWorkDone[]; // была проблема в 56 строки, ты присваивал свойство от undefined
+}
+
+const StudentPage: React.FC<PropsStudentPage> = ({ tasksData, courseData, completedTasksData }) => {
   const [houmPage, setHoumPage] = useState<boolean>(true);
   const [CCSubmit, setCCSubmit] = useState<boolean>(false);
 
@@ -53,7 +63,7 @@ const StudentPage: React.FC = () => {
           {houmPage ? (
             <StudentHome />
           ) : CCSubmit ? (
-            <CrossCheckSubmitPage />
+            <CrossCheckSubmitPage tasksData={tasksData}, courseData={courseData}, completedTasksData={completedTasksData}/>
           ) : (
             <CrossCheckReviewPage />
           )}
@@ -61,6 +71,46 @@ const StudentPage: React.FC = () => {
       </main>
     </MainLayout>
   );
+};
+export const getServerSideProps = async () => {
+  let tasksData: ITask[] = [] as ITask[];
+  await db
+    .collection('TasksArray')
+    .get()
+    .then((snap) => {
+      if (snap !== undefined && snap !== null) {
+        tasksData = snap.docs.map((doc) => doc.data()) as ITask[];
+      } else {
+        tasksData = [] as ITask[];
+      }
+    });
+  let courseData: ICourse[] = [] as ICourse[];
+  await db
+    .collection('sessions')
+    .get()
+    .then((snap) => {
+      if (snap !== undefined && snap !== null) {
+        courseData = snap.docs.map((doc) => doc.data()) as ICourse[];
+      } else {
+        courseData = [] as ICourse[];
+      }
+    });
+
+  let completedTasksData: IWorkDone[] = [] as IWorkDone[];
+  await db
+    .collection('completed_tasks')
+    .get()
+    .then((snap) => {
+      if (snap !== undefined && snap !== null) {
+        completedTasksData = snap.docs.map((doc) => doc.data()) as IWorkDone[];
+      } else {
+        completedTasksData = [] as IWorkDone[];
+      }
+    });
+
+  return {
+    props: { tasksData, courseData, completedTasksData },
+  };
 };
 
 export default StudentPage;
